pyproject.toml
[project]
name = "tiny-orm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "mysql-connector==2.2.9",
    "mysql-connector-python>=9.6.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv==1.2.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
